+++ src/compute_cache.cc
@@ -116,15 +116,17 @@ void compute_cache::show(FILE *s, bool verbose) const
 { 
   char filler[] = "\t";
   fprintf(s, "%sNumber of slots:\t%d\n", filler, nodeCount);
-  fprintf(s, "%sMemory usage:   \t%u\n", filler,
-      dataCount * sizeof(int) +
-      nodeCount * sizeof(cache_entry) +
-      (ht == 0? 0: ht->getMemoryUsage()) +
-      (fsht == 0? 0: fsht->getMemoryUsage()));
+  fprintf(s, "%sMemory usage:   \t%lu\n",
+      filler, (unsigned long)(
+        dataCount * sizeof(int) +
+        nodeCount * sizeof(cache_entry) +
+        (ht == 0? 0: ht->getMemoryUsage()) +
+        (fsht == 0? 0: fsht->getMemoryUsage())));
   if (verbose) {
-    fprintf(s, "%s  Nodes[]:      \t%u\n",
-        filler, nodeCount * sizeof(cache_entry));
-    fprintf(s, "%s  Data[]:       \t%u\n", filler, dataCount * sizeof(int));
+    fprintf(s, "%s  Nodes[]:      \t%lu\n",
+        filler, (unsigned long)(nodeCount * sizeof(cache_entry)));
+    fprintf(s, "%s  Data[]:       \t%lu\n",
+        filler, (unsigned long)(dataCount * sizeof(int)));
   }
   fprintf(s, "%sPings:          \t%d\n", filler, pings);
   fprintf(s, "%sHits:           \t%d\n", filler, hits);
